Abstracted Files
Split up analysis page into frontend components, working but needs styling work.
